# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Hop Query API
  description: |
    Point an agent at this spec (or /llms.txt, /SKILL.md). Hedera exact x402
    in seconds: unpaid POST /v1/query returns 402, retry with X-PAYMENT. No
    API key. Human docs: /docs. Swagger: /swagger.html. Verifiable decision
    infrastructure for agents. Demo vertical: paid lending policy queries.
    Settlement is public Hedera x402 `exact` via Blocky402.
    Data is Messari Lending/CDP schema 3.1.0 from The Graph Gateway. The join
    runs in a Chainlink CRE `handlerInTee` workflow (simulation unless a DON
    result is retrieved and matched). Query types are enumerated; Hop does not
    parse natural language. Policy values are omitted from the result.
    Technical reference: documentation/README.md. Machine files also at
    /llms.txt and /SKILL.md.
  version: 0.0.70
servers:
  - url: http://localhost:8787
paths:
...
  /v1/query:
    post:
      operationId: query
      summary: Paid decision query (finance demo)
      parameters:
        - name: X-PAYMENT
          in: header
          required: false
          description: Base64 x402 v2 PaymentPayload on the paid retry. PAYMENT-SIGNATURE is also accepted.
          schema:
            type: string
        - name: Idempotency-Key
          in: header
          required: false
          description: Required on paid retry. Same key → same evidence id, no second settle.
          schema:
            type: string
        - name: X-Hop-Mandate
          in: header
          required: false
          description: Mandate JSON or base64. Budget, velocity, expiry. LLM never pays.
          schema:
            type: string
        - name: X-Hop-Confirm
          in: header
          required: false
          description: "1 after mandate_review (human threshold)."
          schema:
            type: string
        - name: X-Hop-World
          in: header
          required: false
          description: World ID session token from POST /v1/world/verify. Required when WORLD_REQUIRED=1.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryRequest"
      responses:
        "200":
          description: |
            Charged outcome. Settlement happens before the join (charge-for-attempt).
            `accept`, `reject`, `stale`, and `k_anon_denied` are all settled.
            `stale` means the join was unevaluable; persist `receipt.evidence_id` and do not reuse the payment payload.
            Idempotent replay returns the same evidence with `receipt.charge.semantics=idempotent_replay` and is not settled again.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryResponse"
        "400":
          description: Bad 402 payload — not charged
        "402":
          description: Payment required — not charged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentRequired"
        "403":
          description: payer_denied, world_required, or mandate DENY/REVIEW — not charged
        "409":
          description: In-flight idempotent request ended before evidence persistence — not charged again
        "429":
          description: rate_limited — not charged
        "503":
          description: policy_unavailable | graph_unconfigured | cre_unavailable | facilitator_unavailable | merchant_unconfigured — not charged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorBody"
  /v1/evidence:
    get:
      operationId: listEvidence
      summary: Recent public evidence packs, newest first
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
      responses:
        "200":
          description: Evidence list
          content:
            application/json:
              schema:
                type: object
                required: [items, count]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Evidence"
                  count:
                    type: integer
  /v1/evidence/{id}:
    get:
      operationId: getEvidence
      summary: Public evidence pack (hashes, no rows, no caps)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Evidence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "404":
          description: Unknown id
...
  /v1/mandate:
    get:
      operationId: mandate
      summary: Remaining mandate budget and velocity. Does not pay.
      parameters:
        - name: X-Hop-Mandate
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Mandate inspect
        "404":
          description: mandate_missing
  /v1/evidence/{id}/peac:
    get:
      operationId: getPeac
      summary: PEAC-shaped portable receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PEAC-shaped JSON
        "404":
          description: Unknown id
  /v1/evidence/{id}/verify:
    get:
      operationId: verifyEvidence
      summary: Recompute hashes and confirm public Hedera settlement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: refresh
          in: query
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: |
            `ok` is local hashes + public settlement (+ HCS if present) + structural CRE.
            `cre_ok` is structural CRE fields only. `tiers.cre_don_verified` is true only when
            `cre.mode` is `don` after a commitment-matched DON execution. `tiers.hcs_confirmed`
            is vacuously true when `hcs_present` is false. `receipt` is `hop.decision.v1`.
          content:
            application/json:
              schema:
                type: object
                required: [ok, tiers, ok_means, cre_ok_means, hcs_present]
                properties:
                  ok: { type: boolean }
                  ok_means:
                    type: string
                    const: local_hashes_and_public_settlement
                  cre_ok_means:
                    type: string
                    const: structural_simulation_fields
                  hcs_present: { type: boolean }
                  tiers:
                    type: object
                    required: [recomputed, settlement_confirmed, hcs_confirmed, cre_simulation, cre_don_verified]
                    properties:
                      recomputed: { type: boolean }
                      settlement_confirmed: { type: boolean }
                      hcs_confirmed: { type: boolean }
                      cre_simulation: { type: boolean }
                      cre_don_verified: { type: boolean }
                  receipt:
                    $ref: "#/components/schemas/VerifiableDecisionReceipt"
                  aggregate_ok: { type: boolean }
                  chain_ok: { type: boolean }
                  predecessor_ok: { type: boolean }
                  peac_ok: { type: boolean }
                  cre_ok: { type: boolean }
                  settlement_ref_present: { type: boolean }
                  external_settlement_verified: { type: boolean }
                  external_hcs_verified: { type: boolean }
        "404":
          description: Unknown id
...
  /v1/demo/sign:
    post:
      operationId: demoSign
      summary: Opt-in ExactHederaScheme signer (HOP_DEMO_SIGN=1)
      responses:
        "200":
          description: payment base64
        "403":
          description: demo_sign_disabled
...
components:
  schemas:
    QueryType:
      type: string
      enum:
        - market_params
        - position_counts
        - liquidations
        - policy_check
        - account_ltv
    QueryRequest:
      type: object
      required: [query, protocols, max_block_lag]
      properties:
        query:
          $ref: "#/components/schemas/QueryType"
        protocols:
          type: array
          description: One or two configured protocol keys or slugs from GET /v1/meta. The caller agent chooses these sources.
          items:
            type: string
          minItems: 1
          maxItems: 2
        max_block_lag:
          type: integer
          minimum: 0
        window:
          type: object
          description: liquidations only
          properties:
            from:
              type: string
              format: date-time
            to:
              type: string
              format: date-time
    QueryStatus:
      type: string
      enum:
        - accept
        - reject
        - k_anon_denied
        - stale
    QueryResponse:
      type: object
      required: [status, evidence, receipt]
      properties:
        status:
          $ref: "#/components/schemas/QueryStatus"
        aggregate:
          description: Omitted on stale and k_anon_denied
          type: object
          additionalProperties: true
        evidence:
          $ref: "#/components/schemas/Evidence"
        receipt:
          $ref: "#/components/schemas/VerifiableDecisionReceipt"
        reason:
          type: object
        mandate:
          type: object
        trace:
          type: array
    VerifiableDecisionReceipt:
      type: object
      required: [schema, decision, privacy, payment, charge, graph, cre, hashes, evidence_id]
      properties:
        schema:
          type: string
          const: hop.decision.v1
        decision:
          type: object
          required: [status, query, demo_vertical]
          properties:
            status:
              $ref: "#/components/schemas/QueryStatus"
            query:
              $ref: "#/components/schemas/QueryType"
            demo_vertical:
              type: string
              const: finance.lending_policy_gate
        privacy:
          type: object
          required: [policy_values, graph_source, settlement]
          properties:
            policy_values:
              type: string
              const: omitted
            graph_source:
              type: string
              const: public
            settlement:
              type: string
              const: public
        payment:
          type: object
          required: [network, scheme, ref, amount]
          properties:
            network:
              type: string
              const: hedera:testnet
            scheme:
              type: string
              const: exact
            ref:
              type: string
            payer_account:
              type: string
            amount:
              type: string
        charge:
          type: object
          required: [semantics, settled]
          properties:
            semantics:
              type: string
              enum: [attempt, idempotent_replay]
            settled:
              type: boolean
        graph:
          type: object
        cre:
          type: object
        hashes:
          type: object
          required: [aggregate, policy]
          properties:
            aggregate: { type: string }
            policy: { type: string }
            peac: { type: string }
            chain: { type: string }
            cre_commitment: { type: string }
        evidence_id:
          type: string
        verification:
          type: object
          description: Omitted on POST /v1/query. Filled by GET /v1/evidence/{id}/verify.
    PaymentRequired:
      type: object
      required: [x402Version, accepts]
      properties:
        x402Version:
          type: integer
          example: 2
        accepts:
          type: array
          items:
            $ref: "#/components/schemas/PaymentRequirements"
    PaymentRequirements:
      type: object
      required: [scheme, network, amount, payTo, maxTimeoutSeconds, asset, extra]
      properties:
        scheme:
          type: string
          enum: [exact]
        network:
          type: string
          example: hedera:testnet
        amount:
          type: string
          description: tinybars when asset is 0.0.0
        payTo:
          type: string
        maxTimeoutSeconds:
          type: integer
        asset:
          type: string
          description: "0.0.0 for HBAR, or HTS token id"
          example: "0.0.0"
        extra:
          type: object
          required: [feePayer]
          properties:
            feePayer:
              type: string
              description: MUST match facilitator GET /supported
    ErrorBody:
      type: object
      required: [error]
      properties:
        error:
          type: string
          enum:
            - policy_unavailable
            - not_implemented
            - bad_payment
            - idempotency_required
            - idempotency_conflict
            - idempotency_incomplete
            - payer_denied
            - rate_limited
            - not_found
            - payment_replay
            - graph_unconfigured
            - payload_too_large
            - facilitator_unavailable
            - merchant_unconfigured
            - mandate_required
            - mandate_expired
            - mandate_merchant
            - mandate_query
            - mandate_freshness
            - mandate_per_call
            - mandate_budget
            - mandate_velocity
            - world_required
            - world_unconfigured
            - mandate_review
            - bad_mandate
    Evidence:
      type: object
      required:
        - id
        - timestamp
        - query
        - graph
        - policy
        - k_anon
        - aggregate_hash
        - settlement
        - cre
        - status
      properties:
        id:
          type: string
        timestamp:
          type: string
          format: date-time
        payer_account:
          type: string
        query:
          type: object
          required: [type, params]
          properties:
            type:
              $ref: "#/components/schemas/QueryType"
            params:
              type: object
              additionalProperties: true
              description: No wallet list
        graph:
          type: object
          required: [deployments]
          properties:
            deployments:
              type: array
              items:
                type: object
                required: [id, subgraphId, schemaVersion]
                properties:
                  id:
                    type: string
                    deprecated: true
                    description: Legacy alias for subgraphId (Gateway Subgraph ID).
                  subgraphId:
                    type: string
                    description: Graph Gateway / Studio Subgraph ID used in /subgraphs/id/{id}.
                  deploymentId:
                    type: string
                    description: Graph `_meta.deployment` (deployment/IPFS identifier) when returned by the Gateway.
                  slug:
                    type: string
                  schemaVersion:
                    type: string
                    example: "3.1.0"
                  subgraphVersion:
                    type: string
                  methodologyVersion:
                    type: string
                    description: Messari methodology version from lendingProtocols.
                  block:
                    type: integer
                  blockTimestamp:
                    type: integer
        policy:
          type: object
          required: [version, threshold_hash]
          properties:
            version:
              type: string
            threshold_hash:
              type: string
              description: Hash of private table. Never the caps.
        k_anon:
          type: object
          required: [result]
          properties:
            result:
              type: string
              enum: [pass, fail, not_applicable]
        aggregate_hash:
          type: string
        settlement:
          type: object
          required: [ref]
          properties:
            ref:
              type: string
              description: Blocky402 settlement ref
        cre:
          type: object
          required: [mode]
          properties:
            mode:
              type: string
              enum: [simulation, don]
            artifact:
              type: string
            tee:
              type: string
            trigger:
              type: string
            cre_commitment_hash:
              type: string
              description: SHA-256 of the JSON commitment passed to usingTheDons().report().
            report_hash:
              type: string
              deprecated: true
              description: Historical alias for cre_commitment_hash.
            execution_id:
              type: string
              description: CRE workflow_execution_id when a DON trigger was accepted. Not proof of a matched result.
            don_status:
              type: string
              description: CRE execution status from `cre execution status` when retrieved.
            don_executed_in_tee:
              type: boolean
        world:
          type: object
          properties:
            nullifier_hash:
              type: string
              description: SHA-256 of World ID nullifier. Not the raw nullifier.
        status:
          $ref: "#/components/schemas/QueryStatus"
        hcs_seq:
          type: integer
          description: HCS topic sequence when operator (and topic or HOP_HCS_AUTO) is set
        mandate:
          type: object
          properties:
            id:
              type: string
            hash:
              type: string
            remaining_tinybars:
              type: integer
            remaining_hops:
              type: integer
            decision:
              type: string
              enum: [ALLOW, DENY, REVIEW]
        chain:
          type: object
          properties:
            prev:
              type: string
            hash:
              type: string
        meter:
          type: object
          properties:
            amount:
              type: string
            protocols:
              type: integer
        peac_hash:
          type: string
        reason:
          type: object
          properties:
            stamp:
              type: string
            metric_hash:
              type: string
            observed:
              type: number
            freshness:
              type: string
              enum: [live, stale]
        posture:
          type: object
          properties:
            custody:
              type: string
            ofac:
              type: string
            mor:
              type: string
            cre:
              type: string
            ats:
              type: string